requirements.txt
python-dateutil~=2.8.1
pytz~=2020.1
rstr~=2.2.6
prettytable~=0.7.2
setuptools~=49.6.0
coverage~=5.2.1